Det virker (noget af det)

+++ I4DAB/Handin2_1/Handin2_1/AA.cs
@@ -7,7 +7,7 @@ namespace Handin2_1
 {
     public class AA
     {
-        public virtual int _AAID { get; set; }
+        public virtual long _AAID { get; set; }
         public virtual Person TilknyttetPerson { get; set; }
         public virtual Adresse TilknyttetAdresse { get; set; }
         public virtual string _AdresseType { get; set; }
