+++ I4DAB/Handin2_1/Handin2_1/Email.cs
@@ -9,6 +9,6 @@ namespace Handin2_1
     public class Email
     {
         public virtual string _EmailAdresse { get; set; }
-        public virtual int _EmailID { get; set; }
+        public virtual long _EmailID { get; set; }
     }
 }
